+++ plugins/MET.cc
@@ -7,11 +7,11 @@ BristolNTuple_MET::BristolNTuple_MET(const edm::ParameterSet& iConfig) :
 		inputTag(iConfig.getParameter < edm::InputTag > ("InputTag")), //
 		prefix(iConfig.getParameter < std::string > ("Prefix")), //
 		suffix(iConfig.getParameter < std::string > ("Suffix")) {
-	produces<double>(prefix + "Ex" + suffix);
-	produces<double>(prefix + "Ey" + suffix);
-	produces<double>(prefix + "ET" + suffix);
-	produces<double>(prefix + "Phi" + suffix);
-	produces<double>(prefix + "Significance" + suffix);
+	produces<float>(prefix + "Ex" + suffix);
+	produces<float>(prefix + "Ey" + suffix);
+	produces<float>(prefix + "ET" + suffix);
+	produces<float>(prefix + "Phi" + suffix);
+	produces<float>(prefix + "Significance" + suffix);
 }
 
 void BristolNTuple_MET::produce(edm::Event& iEvent, const edm::EventSetup& iSetup) {
@@ -23,11 +23,11 @@ void BristolNTuple_MET::produce(edm::Event& iEvent, const edm::EventSetup& iSetu
 
 	const pat::MET patMET(mets->at(0));
 
-	std::auto_ptr<double> px(new double(patMET.px()));
-	std::auto_ptr<double> py(new double(patMET.py()));
-	std::auto_ptr<double> met(new double(patMET.pt()));
-	std::auto_ptr<double> phi(new double(patMET.phi()));
-	std::auto_ptr<double> significance(new double(patMET.significance()));
+	std::auto_ptr<float> px(new float(patMET.px()));
+	std::auto_ptr<float> py(new float(patMET.py()));
+	std::auto_ptr<float> met(new float(patMET.pt()));
+	std::auto_ptr<float> phi(new float(patMET.phi()));
+	std::auto_ptr<float> significance(new float(patMET.significance()));
 
 	iEvent.put(px, prefix + "Ex" + suffix);
 	iEvent.put(py, prefix + "Ey" + suffix);
